# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CasaRoca System AI · API
  version: "1.0.0"
  description: |
    API del sistema eclesial de Casa Sobre la Roca: una central y 36 sedes.

    ESTE ARCHIVO SE GENERA. No lo edite a mano: corra `scripts/generar-openapi.js`.
    La verificacion de entrega compara lo generado con lo versionado.

    Autenticacion: token de portador (Bearer). El token dice que sesion DICE ser;
    quien decide es la base, que en cada peticion responde si esa sesion sigue viva.
    Por eso cerrar una sesion surte efecto en el instante.
servers:
  - url: https://api.casaroca.org
    description: produccion
  - url: https://api.staging.casaroca.org
    description: pruebas
  - url: http://127.0.0.1:3000
    description: desarrollo
components:
  securitySchemes:
    portador:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: boolean }
        mensaje: { type: string, description: "En castellano y con la accion siguiente. Nunca filtra la estructura interna." }
        peticionId: { type: string, description: "Identificador de traza. El usuario lo lee por telefono al soporte." }
security:
  - portador: []
paths:
...
  /api/v1/cumplimiento/consentimientos/:personaId:
    get:
      summary: "Que consentimientos tiene hoy una persona"
      description: "Canal por canal y finalidad por finalidad, con su base legal, mas la historia completa de actos."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/cumplimiento/consentimientos/revocar:
    post:
      summary: "Revocar un consentimiento"
      description: "Sin canal ni finalidad revoca todo lo REVOCABLE. Lo que se apoya en contrato u obligacion legal no se revoca: decir que si seria mentirle al titular sobre su propio derecho. Ademas descarta lo que ya estaba encolado."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/cumplimiento/no-atendido:
    get:
      summary: "Lo que el sistema decidio NO enviar"
      description: "Con su motivo. Es la prueba de que no se contacto a quien no autorizo, y la bandeja de lo que quedo sin atender."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/cumplimiento/peticiones/:id/prorrogar:
    post:
      summary: "Prorrogar el plazo"
      description: "La base exige un motivo de verdad y recalcula el vencimiento. Avisa si la prorroga todavia NO se le ha informado al titular, que es lo que obliga la ley."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/cumplimiento/peticiones/:id/responder:
    post:
      summary: "Responder y cerrar"
      description: "Deja la respuesta escrita y avisa si se respondio DESPUES del vencimiento. Una respuesta tardia sigue siendo un incumplimiento y el registro tiene que poder demostrarlo."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/cumplimiento/peticiones/:id/suprimir:
    post:
      summary: "Ejecutar la supresion"
      description: "Irreversible. Exige confirmacion explicita (\"SUPRIMIR\") y nivel N4, porque puede tocar datos de menores y de consejeria."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/cumplimiento/peticiones:
    get:
      summary: "La bandeja de peticiones"
      description: "Lo abierto primero y lo vencido arriba. Si hay vencidas lo dice en un aviso: una peticion vencida es un incumplimiento en curso, no un dato de una columna."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      summary: "Radicar una peticion del titular"
      description: "Consulta, reclamo, supresion, revocacion o actualizacion. La base calcula el vencimiento en DIAS HABILES con los festivos de Colombia y devuelve el radicado. Exige N3: una peticion trae nombre, documento y contacto de una persona."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...